Refactored design of library management system

+++ src/main/java/solutions/librarymanagementsystem/services/BookService.java
@@ -1,10 +1,12 @@
 package solutions.librarymanagementsystem.services;
 
+import solutions.librarymanagementsystem.exception.LMSException;
 import solutions.librarymanagementsystem.models.Book;
 import solutions.librarymanagementsystem.models.BookItem;
 import solutions.librarymanagementsystem.repositories.BookRepository;
 
 import java.util.List;
+import java.util.Optional;
 import java.util.stream.Collectors;
 
 public class BookService {
@@ -16,11 +18,27 @@ public class BookService {
     }
 
     public void addBookItem(BookItem bookItem) {
-        this.bookRepository.addBookItem(bookItem);
+        String bookItemId = bookItem.getBookItemId();
+        Optional<BookItem> bookItemOptional = Optional.ofNullable(this.bookRepository.getBookItem(bookItemId));
+
+        if (bookItemOptional.isEmpty()) {
+            this.bookRepository.addBookItem(bookItem);
+        } else {
+            throw new LMSException("Book item already exists in the library.");
+        }
     }
 
     public void removeBookItem(BookItem bookItem) {
-        this.bookRepository.removeBookItem(bookItem);
+
+        String bookItemId = bookItem.getBookItemId();
+        Optional<BookItem> bookItemOptional = Optional.ofNullable(this.bookRepository.getBookItem(bookItemId));
+
+        if (bookItemOptional.isPresent()) {
+            this.bookRepository.removeBookItem(bookItem);
+        } else {
+            throw new LMSException("Book item doesnt exist in the library.");
+        }
+
     }
 
     public BookItem getBookItem(String bookItemId) {
@@ -28,9 +46,22 @@ public class BookService {
     }
 
     public List<Book> searchBooksByTitle(String keyword) {
-        return this.bookRepository.getAllBooks().stream().filter(book ->
-                book.getTitle().contains(keyword)
-        ).collect(Collectors.toList());
+        return this.bookRepository
+                .getAllBooks()
+                .stream()
+                .filter(book -> book.getTitle().contains(keyword))
+                .collect(Collectors.toList());
+    }
+
+    public List<BookItem> getAllBookItemsByBook(Book book) {
+
+        String bookId = book.getId();
+        Optional<Book> bookOptional = Optional.ofNullable(this.bookRepository.getBook(bookId));
+
+        if (bookOptional.isEmpty()) {
+            throw new LMSException("Book doesnt exist in the library.");
+        }
+        return this.bookRepository.getAllBookItemsByBook(book);
     }
