+++ src/main/java/solutions/librarymanagementsystem/services/MemberService.java
@@ -1,7 +1,11 @@
 package solutions.librarymanagementsystem.services;
 
+import solutions.librarymanagementsystem.exception.LMSException;
+import solutions.librarymanagementsystem.models.BookItem;
 import solutions.librarymanagementsystem.models.Member;
 import solutions.librarymanagementsystem.repositories.MemberRepository;
+
+import java.util.Optional;
 
 public class MemberService {
 
@@ -12,11 +16,26 @@ public class MemberService {
     }
 
     public void addMember(Member member) {
-        this.memberRepository.addMember(member);
+
+        String memberId = member.getMemberId();
+        Optional<Member> memberOptional = Optional.ofNullable(this.memberRepository.getMember(memberId));
+
+        if (memberOptional.isEmpty()) {
+            this.memberRepository.addMember(member);
+        } else {
+            throw new LMSException("Member with id " + memberId + " already exists in the system");
+        }
     }
 
     public void removeMember(String memberId) {
-        this.memberRepository.removeMember(memberId);
+
+        Optional<Member> memberOptional = Optional.ofNullable(this.memberRepository.getMember(memberId));
+
+        if (memberOptional.isPresent()) {
+            this.memberRepository.removeMember(memberId);
+        } else {
+            throw new LMSException("Member doesn't exist in the library.");
+        }
     }
 
     public Member getMember(String memberId) {
